added edit user option

+++ src/components/StoreItem.js
@@ -1,6 +1,17 @@
-import React, { useState } from "react";
-import { Card, Button, Modal, Container, Row, Col } from "react-bootstrap";
+import React, { useContext, useEffect, useState } from "react";
+import {
+  Card,
+  Button,
+  Modal,
+  Container,
+  Row,
+  Col,
+  Form,
+  InputGroup,
+} from "react-bootstrap";
 import { projectFirestore, projectStorage } from "../firebase/config";
+import { GlobalContext } from "../globalState/globalState";
+import EditProgress from "./EditProgress";
 
 const StoreItem = ({
   id,
@@ -14,10 +25,61 @@ const StoreItem = ({
   addCart,
   setSuccess,
 }) => {
+  const [state, setState] = useContext(GlobalContext);
   const [openDeleteMessage, setOpenDeleteMessage] = useState(false);
+  const [openEditMessage, setOpenEditMessage] = useState(false);
   const [loading, setLoading] = useState(false);
+  const [file, setFile] = useState(null);
+  const [error, setError] = useState(null);
+  const [editSuccess, setEditSuccess] = useState(false);
+  const [originalImage, setOriginalImage] = useState(null);
+
+  const listingsCollection = "listings";
+  const types = ["image/png", "image/jpeg"];
+  const editItemForm = document.getElementById("editItemForm");
 
   const listingsRef = projectFirestore.collection("listings");
+
+  useEffect(() => {
+    if (state.editRun) {
+      const imageRef =
+        originalImage && projectStorage.refFromURL(originalImage);
+      listingsRef
+        .doc(state.editFormData.id)
+        .set(state.editFormData, { merge: true })
+        .then(() => {
+          if (originalImage) {
+            imageRef.delete().then(() => {
+              setState({
+                editRun: false,
+                editLoad: false,
+                editFormData: {},
+              });
+              setEditSuccess(true);
+              setTimeout(() => {
+                setEditSuccess(false);
+                handleCloseModal("edit");
+              }, 2000);
+            });
+          } else {
+            setState({
+              editRun: false,
+              editLoad: false,
+              editFormData: {},
+            });
+            setEditSuccess(true);
+            setTimeout(() => {
+              setEditSuccess(false);
+              handleCloseModal("edit");
+            }, 2000);
+          }
+        })
+        .catch((err) => console.error(err));
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [state.editRun]);
+
+  console.log(originalImage);
 
   const handleDeleteItem = (e, itemId, url) => {
     const imageRef = projectStorage.refFromURL(url);
@@ -27,21 +89,32 @@ const StoreItem = ({
       .doc(itemId)
       .delete()
       .then(() => {
-        imageRef
-          .delete()
-          .then(() => {
-            setSuccess(true);
-            setTimeout(() => {
-              setSuccess(false);
+        if (url) {
+          imageRef
+            .delete()
+            .then(() => {
+              setSuccess(true);
               setLoading(false);
               setOpenDeleteMessage(false);
-            }, 2500);
+              setTimeout(() => {
+                setSuccess(false);
+              }, 2500);
 
-            console.log("Image successfully deleted");
-          })
-          .catch((error) => {
-            console.error("Error removing image: ", error);
-          });
+              console.log("Image successfully deleted");
+            })
+            .catch((error) => {
+              console.error("Error removing image: ", error);
+            });
+        } else {
+          setSuccess(true);
+          setTimeout(() => {
+            setSuccess(false);
+            setLoading(false);
+            setOpenDeleteMessage(false);
+          }, 2500);
+
+          console.log("Image successfully deleted");
+        }
         console.log("Document successfully deleted!");
       })
       .catch((error) => {
@@ -49,14 +122,75 @@ const StoreItem = ({
       });
   };
 
-  const handleOpenModal = (e) => {
-    e.preventDefault();
-    setOpenDeleteMessage(true);
+  const handleOpenModal = (messageType) => {
+    if (messageType === "edit") {
+      setOpenEditMessage(true);
+    }
+    if (messageType === "delete") {
+      setOpenDeleteMessage(true);
+    }
   };
 
-  const handleCloseModal = (e) => {
+  const handleCloseModal = (messageType) => {
+    if (messageType === "edit") {
+      setOpenEditMessage(false);
+    }
+    if (messageType === "delete") {
+      setOpenDeleteMessage(false);
+    }
+  };
+
+  const handleEditFormChange = (e) => {
+    const { name, value } = e.target;
+    setState({
+      editFormData: {
+        ...state.editFormData,
+        [name]: value,
+      },
+    });
+  };
+
+  const handleEditImageChange = (e) => {
+    let selected = e.target.files[0];
+    if (selected && types.includes(selected.type)) {
+      setFile(selected);
+      setError("");
+    } else {
+      setFile(null);
+      setError("Please select an image file (png or jpeg)");
+    }
+  };
+
+  const handleEditFormSubmit = (e, itemId, url) => {
     e.preventDefault();
-    setOpenDeleteMessage(false);
+    if (file) {
+      setOriginalImage(url);
+      setState({
+        editLoad: true,
+        editFormData: {
+          ...state.editFormData,
+          id: itemId,
+        },
+      });
+    } else {
+      listingsRef
+        .doc(itemId)
+        .set(state.editFormData, { merge: true })
+        .then(() => {
+          setState({
+            editRun: false,
+            editLoad: false,
+            editFormData: {},
+          });
+          editItemForm.reset();
+          setEditSuccess(true);
+          setTimeout(() => {
+            setEditSuccess(false);
+            handleCloseModal("edit");
+          }, 2000);
+        })
+        .catch((err) => console.error(err));
+    }
   };
 
   return (
@@ -80,22 +214,33 @@ const StoreItem = ({
             </Button>
           )}{" "}
           {editItem && (
-            <Button size="sm" variant="secondary">
+            <Button
+              onClick={() => handleOpenModal("edit")}
+              size="sm"
+              variant="secondary"
+            >
               Edit Item
             </Button>
           )}{" "}
           {deleteItem && (
-            <Button onClick={handleOpenModal} size="sm" variant="warning">
+            <Button
+              onClick={() => handleOpenModal("delete")}
+              size="sm"
+              variant="warning"
+            >
               Delete Item
             </Button>
           )}
         </Card.Body>
       </Card>
       <Modal show={openDeleteMessage}>
-        <Modal.Header closeButton></Modal.Header>
+        <Modal.Header
+          closeButton
+          onHide={() => handleCloseModal("delete")}
+        ></Modal.Header>
         <Container className="mt-5 mb-5">
-          <Row>
-            <Col className="text-center">
+          <Row className="align-items-center">
+            <Col className="text-center pb-3">
               <h4>Are you sure you want to delete this item?</h4>
             </Col>
           </Row>
@@ -110,7 +255,7 @@ const StoreItem = ({
                 <Button
                   size="sm"
                   variant="secondary"
-                  onClick={handleCloseModal}
+                  onClick={() => handleCloseModal("delete")}
                 >
                   Cancel
                 </Button>{" "}
@@ -123,6 +268,126 @@ const StoreItem = ({
                 </Button>
               </Col>
             )}
+          </Row>
+        </Container>
+      </Modal>
+      <Modal show={openEditMessage} size="xl">
+        <Modal.Header closeButton onHide={() => handleCloseModal("edit")}>
+          <h4>Edit Item Details</h4>
+        </Modal.Header>
+        <Container className="mt-3 mb-3">
+          <Row>
+            <Col>
+              <Form
+                id="editItemForm"
+                onSubmit={(e) => handleEditFormSubmit(e, id, photo)}
+              >
+                <Row>
+                  <Col>
+                    <Form.Group>
+                      <Form.Label>Item Name</Form.Label>
+                      <Form.Control
+                        onChange={handleEditFormChange}
+                        type="text"
+                        defaultValue={itemName}
+                        name="name"
+                      />
+                    </Form.Group>
+                  </Col>
+                </Row>
+                <Row>
+                  <Col>
+                    <Form.Group>
+                      <Form.Label>Item Description</Form.Label>
+                      <Form.Control
+                        defaultValue={itemDescription}
+                        onChange={handleEditFormChange}
+                        as="textarea"
+                        type="text"
+                        name="description"
+                      />
+                    </Form.Group>
+                  </Col>
+                </Row>
+                <Row>
+                  <Col>
+                    <Form.Label>Price</Form.Label>
+                    <InputGroup>
+                      <InputGroup.Prepend>
+                        <InputGroup.Text>$</InputGroup.Text>
+                      </InputGroup.Prepend>
+
+                      <Form.Control
+                        onChange={handleEditFormChange}
+                        name="price"
+                        type="text"
+                        required
+                        placeholder="0.00"
+                        defaultValue={price}
+                      />
+                    </InputGroup>
+                  </Col>
+                  <Col>
+                    <Form.Group>
+                      <Form.Label>Quantity</Form.Label>
+                      <Form.Control
+                        onChange={handleEditFormChange}
+                        type="number"
+                        name="quantity"
+                        defaultValue={quantity}
+                      />
+                    </Form.Group>
+                  </Col>
+                  <Col>
+                    <Form.Group>
+                      <Form.Label>Image</Form.Label>
+                      <Form.Control
+                        onChange={handleEditImageChange}
+                        type="file"
+                        name="url"
+                      />
+                    </Form.Group>
+                    <img src={photo} width="50px" alt={itemName} />
+                  </Col>
+                </Row>
+                <Row>
+                  <Col>
+                    {editSuccess && (
+                      <div className="success">
+                        Item was updated successfully
+                      </div>
+                    )}
+                    {error && <div className="error">{error}</div>}
+                    {state.editLoad && file && (
+                      <EditProgress
+                        file={file}
+                        setFile={setFile}
+                        collection={listingsCollection}
+                      />
+                    )}
+                  </Col>
+                </Row>
+                <Row>
+                  <Col>
+                    <Button
+                      variant="secondary"
+                      size="sm"
+                      onClick={() => handleCloseModal("edit")}
+                    >
+                      Cancel
+                    </Button>{" "}
+                    <Button
+                      disabled={editSuccess}
+                      type="submit"
+                      variant="info"
+                      size="sm"
+                    >
+                      Update
+                    </Button>
+                  </Col>
+                </Row>
+              </Form>
+            </Col>
           </Row>
         </Container>
       </Modal>
